requirements.txt
numpy>=1.24
scikit-image>=0.21
scikit-learn>=1.3
opencv-python>=4.8
Pillow>=10.0
PyQt5>=5.15
rasterio>=1.3
